+++ tools/sense_studio/static/video_recording.js
