+++ tools/sense_studio/templates/frame_annotation.html
