+++ tools/sense_studio/templates/project_details.html
@@ -12,7 +12,7 @@
         </div>
     </h1>
     <a class="ui icon button hashoverpopup" onclick="loading(this);"
-       href="{{ url_for('prepare_annotations_bp.prepare_annotation', path=path) }}"
+       href="{{ url_for('annotations_bp.prepare_annotation', path=path) }}"
        data-content="Prepare required files for all classes">
         <i class="file video outline icon"></i>
         Prepare Annotations
