+++ tools/sense_studio/templates/video_list.html
