+++ tools/sense_studio/templates/video_recording.html
